Adds 'cli' and 'all' to extras_require
People can now install meltano without the cli (as a library). Not
exceedingly useful at the moment but makes sense as the cli gets more
complicated and requires more libraries specific to it

Let's just use `pip install -e '.[all]'` from now on so the docs don't get
out of date and people don't end up with half of the deps. Also I am
seeing people thinking that '.[api]' is referring to the api directory.