test: Run & test against Snowplow Micro under Pytest (#6257)
* Run & test against Snowplow Micro under Pytest

If Docker/docker-compose is not available, tests using the `snowplow`
fixture will be skipped, and tests using the `snowplow_optional` fixture
will receive `None` instead of a `SnowplowMicro` instance.

Currently all CLI tests that invoke a command are checked to ensure that
no bad events are fired. We also check that an exit event is fired with
an appropriate exit code in one case. In the future more tests can be
added that check that the events fired have the expected events with
their expected contexts.

Closes https://github.com/meltano/meltano/issues/2955

* Update Poetry lock file

* Fix failing tests

* Address lint issues

* Add section about testing with Snowplow Micro to telemetry docs

* Update Snowplow Micro under Pytest docs

* Add contributing tests doc & misc doc cleanup

* Fix autogenerated Markdown table of contents

* Explicitly list supported Python versions in `getting-started.md`

* Fix Pytest "module already imported" warning

* Change Pytest short test summary order

* Remove unused constant from `conftest.py`

* Use monkeypatch to alter env in `snowplow_optional` fixture