chore: Remove Makefile
Several of the commands in the Makefile do not work anymore, are not necessary, or should not be run locally. The Makefile has a considerable amount of dead code within it, and presents a small but significant burden to contributors. It all adds up: the Makefile does not prevent contributors from having to learn various poetry and yarn invocations to build/test/etc. our software - it just adds more on top of that.

From asking around, and searching through the repository, it seems like there are only 2 commonly used `make` commands: `ui` and `bundle`. They have been replaced by `scripts/bundle_ui.sh`, which can be run from anywhere to achieve the same effect.

Besides that, all of the local tasks that used to be carried out through the Makefile should now instead be done through poetry, yarn, and optionally pre-commit.

Much of what was in the Makefile was for CI, which is now all handled in the workflows and actions under `.github/`. In the unlikely circumstance that we need to run those commands locally, due to the simplicity of how we have our CI set up, we can copy-paste the commands from the workflows/actions manually.

Relates to #6262 and #6487