chore: Fix eslint failures in CI
Before running eslint, you have to cd into `src/webapp`. As a result, the pre-commit `eslint` hook was failing when run. We didn't notice this for quite a long while, presumably because the hook was not being run often at all. In #6800 it caused a failure, which prompted me to open #6803.

Making pre-commit run from a subdirectory is somewhat tricky. The maintainer of pre-commit is adament about not supporthing this feature, and suggests that users hack their way around the limitation as I have done here.

I also took the liberty of updating our JS linting plugins to their latest versions. Many of them are duplicated 2 or 3 times across the repository, which is unfortunate, and will surely cause problems in the future (and may already be causing problems now).

I have eslint autofixing where possible. This PR also includes the autofixes that resulting from running `pre-commit run --all-files eslint`. A few manual fixes were applied to the files in `docs/js/`, as well as a few ignores. Some for an undefined variable that was clearly defined, and others for various Vue issues that I am not comfortable fixing at this time.

For the undefined variable there's probably some proper way of making it count as defined, but as-is my editor seems to recognize it as defined, and the code seems to actually work - I'd rather not break it.

Closes #6803