fix: Make default `default_environment` null
Previously if `default_environment` was not set in `meltano.yml` (and the `--environment` CLI option was not used) Meltano would try to use the `dev` environment. Now it is set to null in `settings.yml`, which results in no default environment being used when there isn't one set in `meltano.yml`.

Closes #6833