chore: Set environment context UUID for subprocesses
This PR extends https://github.com/meltano/meltano/pull/7007. With #7007, we can set the environment context UUID using an env var. With this PR, we that env var is automatically set by Meltano when invoking Meltano or a plugin, and is reported using the new `parent_context_uuid` field within version 1-1-0 of the environment context.

Once this PR has been approved, but before it is merged, I will publish the new environment context schema version to Snowcat Cloud.

Meltano can invoke a plugin, which then calls into Meltano (e.g. to list schedules). Now telemetry from the child Meltano process can more easily be related to telemetry from the parent process. This was previously possible using the data from the process tree, but it was difficult.

Most of the changes in this PR were done to make it so that importing the `Tracker` class doesn't import much else. The need to import the tracker in a couple new spots was causing circular imports.

Closes https://github.com/meltano/cloud-planning/issues/85