fix: Ensure `environment_name_hash` is collected
In https://github.com/meltano/meltano/pull/6862 we defered the activate of environments. The `Tracker` was implicitly relying on the environment being activated before it was instantiated, or no environment being activated at all.

Unfortunately fixing that implicit dependency would be more effort than I'm willing to commit to this at this time. Instead, I have made it so that activating an environment will update `environment_name_hash` in the `ProjectContext` in the `Tracker`.

New `ProjectContext` objects created past that point will, as before, get the active environment information from the `Project` object.

I have verified that this works using Snowplow Micro locally. Before this change `environment_name_hash` was always `null`. After this change, `environment_name_hash` is set to the expected hashed string when an environment is active, and `null` when no environment is active. I would encourage at least one reviewer to double check this.

Closes https://github.com/meltano/internal-data/issues/51