test: Catch all exceptions in `snowplow_session` fixture
Context: https://meltano.slack.com/archives/C013Z450LCD/p1670619321698669

This fixture is meant to return `None` if it cannot be started, since it is optional and requires additional dependencies. Previously we were only returning `None` if an exception was raised while checking for `docker-compose`, or while instantiating `SnowplowMicro` itself. This caused many errors during testing for a user for whom Snowplow Micro couldn't find Java.

I don't see why we shouldn't just catch all normal exceptions, to be on the safe side. There is little risk of this resulting in us unknowingly failing to test something, since if `None` is returned then Pytest will skip the tests, and will report that it skipped them.