feat(core): Optionally ignore missing env vars during expansion
In addition to addressing #7122, this also includes some performance improvements that have a significant effect when using `expand_env_vars` on many large env dictionaries, as will be done by the implementation for [Meltano manifest](https://github.com/meltano/meltano/issues/6876).

Some work to wrangle the complexity of `expand_env_vars` was made, but it's widely used by legacy code, which makes it difficult to improve significantly.

For instance, `expand_env_var` could previous be called on any type, and it would return it. This is now disallowed by the type hint, but still supported in the code itself. Hopefully this will discourage these bad practices.

Another questionable aspect of `expand_env_vars` that had to be preserved is that it will perform a recursive tree-traversal on any mapping structure it is provided as `raw_value`. This dangerous/slow/complex behaviour can now be opted out of by setting the `flat` paramter to `True`, which asserts that the `raw_value` being provided has a flat structure (i.e. no nested dicts).

Closes #7122