chore: delete `meltano.core.utils.validate_json_schema`
This appears to be misplaced. It only works when run from the Meltano repository, rather than from an installed copy of Meltano. It is a utility script for developers, rather than something that belongs under `src/meltano/`.

That would give reason to move it into `scripts/`, but instead I am deleting it because it does not appear to do anything more than the equivalent of running `check-jsonschema --schemafile schema/discovery.schema.json src/meltano/core/bundle/discovery.yml`.

Note that the `jsonschema` package/tool can be used instead of `check-jsonschema`, but recent versions of `jsonschema` recommend the use of `check-jsonschema` instead as `jsonschema`'s included CLI is deprecated.

Relates to #7168