chore: Add script to host docs & update Ruby dependencies
This should simplify the process of developing the docs locally because a Ruby installation no longer needs to be installed/managed. It comes at the cost of having to reinstall the Ruby depdencies each time your run `scripts/host_docs.sh`, but that doesn't take long, and the site live-reloads when its content is changed.

Maintaining a Ruby environment outside of Docker and using that manually remains an option. This script offers convenience.

Closes #7195