Add `--indent` option for `meltano compile`
This was mostly added because pre-commit was trying to make the expected manifest files have 2 spaces of indentation, which would make comparing them with `diff` difficult.

We could just have that pre-commit check skip those files, but this is a very easy feature to add, and it seems nice to have.