Fix "must be run inside a Meltano project" error
While editing `tests/meltano/cli/test_cli.py` I also took the liberty of adding 5 seconds to the performance test, since it's a somewhat frequent cause of flakyness in our tests due to the Windows runners running slower. This isn't a good solution (if it can even be considered one), but it's probably worth doing until we have better performance testing in-place.

Closes #7202