chore: Do not run Nox coverage session after tests
When we run `pytest` with Nox, it uses `pytest-cov`, which automatically (and uncontrollably) prints a coverage report. Nox then runs its coverage session, which prints the coverage report a second time. The coverage job is still useful (it's used in CI, among other things), but I don't think it should be run automatically after the tests when running in interactive mode.

Note that this change has no effect on CI, since the coverage session was only being called in interactive mode.