chore: Disable `meltano compile` command
Context:
- https://github.com/meltano/meltano/issues/7306#issuecomment-1432086410

To avoid anyone depending on the manifest feature before we are confident that we can avoid making breaking changes to it, this PR disables the `meltano compile` command. Manifests can continue to be used internally, but once we allow users to generate them via the CLI, and depend on the data they contain, we will have to be careful about what we change.