chore: Remove cyclic dependencies from `meltano.cli`
This approach of using `click.command` with `cli.add_command` is currently used by the Cloud CLI. This PR applies this approach to the rest of Meltano in order to simplify the imports in the `meltano.cli` module. Now there are no more cyclic dependencies, and we can let isort (Ruff) handle the imports.