chore: Use Dependabot version increase strategy
https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#versioning-strategy

Meltano is an application, not a library, so there is little benefit in widening the allowed dependency version ranges. Letting it grow wide makes it difficult to test, so we should keep it as narrow as we can while still ensuring compatibility with all of our other libraries, and with our supported Python versions.