docs: Use small SVG for 404 image
When I hit our 404 page earlier I noticed it took a significant amount of time for the page to load. The culprit seems to be the large 404 JPEG.

I have replaced it with a minified version of the 404 SVG. Note that no 404 SVG existed in the shared Google Drive, so I created one from the 404 AI file with the updated text currently present in the JPEG, and uploaded it to the shared Google Drive.

The old JPEG was 2.2 MB. The new SVG is 321 KB.