chore: Avoid installing `setuptools` and `wheel` in plugin venvs
Relates to:
https://github.com/meltano/meltano/pull/7022

The above PR was made because `singer-python` was broken, and several plugins depended on it. Because `singer-python` was out of our control, instead of fixing it, we added this hack fix into Meltano itself.

The problem with keeping `setuptools` pinned is that it ensures that an old, possibly vulnerable, version of `setuptools` is installed into every plugin venv.

Meltano v3 seems like a good time to rip this bandaid off. Plugins that rely on this old version of `setuptools` should ensure it's installed themselves, using the ability for modern Python packages to install build-time dependencies.