Keep eslint at v8.55.0
I tried to update it to v9.1.0, and it went poorly. There's a new config file format, and even once I sorted that out there's an upstream bug in the `globals` package that we'd have to work around, and then after that there are issues with our plugins, and lint failures.