chore: Enable Ruff `ANN` rules
For this rule I applied the available auto-fixes, which seems to have just added `-> None` to functions that don't have a return statement. Then I used `ruff check . --add-noqa` to ignore the rest of the issues, since addressing them all in one go is impractical. We can chip away at them as we naturally work on parts of the codebase.