chore: Enable Ruff ANN rules (#8678)
* chore: Enable Ruff `FBT` rules

Relates to:
- #7383

* chore: Enable Ruff `ANN` rules

For this rule I applied the available auto-fixes, which seems to have just added `-> None` to functions that don't have a return statement. Then I used `ruff check . --add-noqa` to ignore the rest of the issues, since addressing them all in one go is impractical. We can chip away at them as we naturally work on parts of the codebase.

* Ignore ANN rules for tests

* Address mypy issues

* Apply suggestions from code review

Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>

---------

Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>