chore: Enable Ruff for Alembic migrations (#8680)
* chore: Enable Ruff `FBT` rules

Relates to:
- #7383

* chore: Enable Ruff `ANN` rules

For this rule I applied the available auto-fixes, which seems to have just added `-> None` to functions that don't have a return statement. Then I used `ruff check . --add-noqa` to ignore the rest of the issues, since addressing them all in one go is impractical. We can chip away at them as we naturally work on parts of the codebase.

* Ignore ANN rules for tests

* chore: Enable Ruff `D` rules

Similarly to the `ANN` rules, everything that couldn't be autofixed here was given a `noqa` mark, with the intent being that we'll improve it incrementally as we naturally work on small parts of the code.

* [pre-commit.ci] auto fixes from pre-commit.ci hooks

for more information, see https://pre-commit.ci

* chore: Enable Ruff for Alembic migrations

* Address mypy issues

* Cleanup

* Update pyproject.toml

* Apply suggestions from code review

Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>

* Update src/meltano/cli/initialize.py

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>