feat: Parse relative dates in plugin config (#9119)
<!--

Please, go through these steps when you submit a PR.

1. Make sure your branch is not protected. In particular, avoid making
PRs from the `main` branch of your fork.

2. Give a descriptive title to your PR. We use semantic titles, and the
accepted types and scopes are listed in
https://github.com/meltano/meltano/blob/main/.github/semantic.yml.

   A good title should look like this:

   ```
feat(cli): The `meltano run` command now accepts a `--timeout` option to
limit the time it runs
   ```

3. Provide a description of your changes.

4. Put "Closes #XXXX" in your comment to auto-close the issue that your
PR fixes (if such).

-->

## Description

<!-- Describe the changes introduced by this PR -->

- [x] Decision: make it fully opt-in? This requires the user adding the
`parse_date` value processor and maybe us bulk-updating plugins in the
Hub, which has backwards-compatibility concerns.

Parsing of relative dates for `date_iso8601` settings is the default.
Parsing of relative dates for `string` settings is opt-in with
`parse_date`.

## Related Issues

* https://github.com/meltano/meltano/issues/3347