test: Use `COVERAGE_CORE=sysmon` to speed up line average computation (#9150)
<!--

Please, go through these steps when you submit a PR.

1. Make sure your branch is not protected. In particular, avoid making
PRs from the `main` branch of your fork.

2. Give a descriptive title to your PR. We use semantic titles, and the
accepted types and scopes are listed in
https://github.com/meltano/meltano/blob/main/.github/semantic.yml.

   A good title should look like this:

   ```
feat(cli): The `meltano run` command now accepts a `--timeout` option to
limit the time it runs
   ```

3. Provide a description of your changes.

4. Put "Closes #XXXX" in your comment to auto-close the issue that your
PR fixes (if such).

-->

## Description

<!-- Describe the changes introduced by this PR -->

> In Python 3.12 and above, you can try an experimental core based on
the new
[**`sys.monitoring`**](https://docs.python.org/3/library/sys.monitoring.html#module-sys.monitoring)
module by defining a `COVERAGE_CORE=sysmon` environment variable. This
should be faster, though plugins and dynamic contexts are not yet
supported with it.[^1]

This manages to shave off 30-40 seconds from the test suite, and I think
it'll be even more with CPython 3.14.

## Related Issues

* Closes #XXXX

[^1]:
https://coverage.readthedocs.io/en/latest/cmd.html#execution-coverage-run