fix: Stringify YAML values from top-level `env` key of `meltano.yml` (#9162)
<!--

Please, go through these steps when you submit a PR.

1. Make sure your branch is not protected. In particular, avoid making
PRs from the `main` branch of your fork.

2. Give a descriptive title to your PR. We use semantic titles, and the
accepted types and scopes are listed in
https://github.com/meltano/meltano/blob/main/.github/semantic.yml.

   A good title should look like this:

   ```
feat(cli): The `meltano run` command now accepts a `--timeout` option to
limit the time it runs
   ```

3. Provide a description of your changes.

4. Put "Closes #XXXX" in your comment to auto-close the issue that your
PR fixes (if such).

-->

## Description

<!-- Describe the changes introduced by this PR -->

If a user passes a value in `env` that ruamel.yaml would parse as
anything other than a string, it breaks the subprocess call because it
expects a mapping of `str` to `str` (or also `bytes` to `bytes` on
UNIX).

## Related Issues

* https://github.com/meltano/meltano/pull/9158