fix: Use `config generate` command to generate `airflow.cfg` configuration file (#9208)
**Description**
This PR addresses an issue with generating the airflow.cfg configuration
file. Specifically, it updates the command used to generate the
configuration from airflow --help to airflow config generate, ensuring
that the correct command is executed to generate the configuration.

**Changes include:**

- Replacing the --help argument with config generate to correctly invoke
the configuration generation.

- Redirecting the output from subprocess.PIPE instead of
subprocess.DEVNULL to capture the configuration content.

- Saving the generated configuration to the appropriate file path
(config_file_path).

- Updating the configuration file after it has been generated.

**Related Issues**
Closes https://github.com/meltano/meltano/issues/9177

---------

Co-authored-by: TyvDmytro <dmytro.tyvoniuk@vitechteam.com>
Co-authored-by: Edgar Ramírez Mondragón <edgarrm358@gmail.com>