fix: Avoid writing empty state after an invalid STATE message (#9258)
<!--

Please, go through these steps when you submit a PR.

1. Make sure your branch is not protected. In particular, avoid making
PRs from the `main` branch of your fork.

2. Give a descriptive title to your PR. We use semantic titles, and the
accepted types and scopes are listed in
https://github.com/meltano/meltano/blob/main/.github/semantic.yml.

   A good title should look like this:

   ```
feat(cli): The `meltano run` command now accepts a `--timeout` option to
limit the time it runs
   ```

3. Provide a description of your changes.

4. Put "Closes #XXXX" in your comment to auto-close the issue that your
PR fixes (if such).

-->

## Description

This pull request includes a change to the `writeline` method in
`src/meltano/core/plugin/singer/target.py`. The change ensures that only
valid states are saved, preventing potential overwrites of existing
valid states when the incoming state message is incorrect (which makes
new state empty).

Here is an example of this failure in logs (extraction failed after it
and state was empty):
```
2025-04-21T16:14:46.146989Z [warning  ] Received state is invalid, incremental state has not been updated
2025-04-21T16:14:46.215886Z [info     ] Incremental state has been updated at 2025-04-21 16:14:46.215847+00:00.

+ meltano state get STATE_ID
{"singer_state": {}}
```


<!-- Describe the changes introduced by this PR -->

## Related Issues

* Closes #XXXX

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Edgar Ramírez Mondragón <edgarrm358@gmail.com>