packaging: Restore support for Click 8.1 (#9496)
It's compatible (tested it) with only a few typing caveats, which are
not user-facing.

This is required for the next Meltano to work with Dagster, since it
doesn't allow Click 8.2:

-
https://github.com/dagster-io/dagster/commit/32720dd3790fcde3083c8b84ce6a1fc5368587ed
-
https://github.com/dagster-io/dagster/blame/ef5bd6d0d2149613f56f9a5bc0f2f2a0a9a327a1/python_modules/dagster/setup.py#L84

The alternative would be to wait for Dagster to move up their
constraint:

- https://github.com/dagster-io/dagster/issues/32144

<!--

Please, go through these steps when you submit a PR.

1. Make sure your branch is not protected. In particular, avoid making
PRs from the `main` branch of your fork.

2. Give a descriptive title to your PR. We use semantic titles, and the
accepted types and scopes are listed in
https://github.com/meltano/meltano/blob/main/.github/semantic.yml.

   A good title should look like this:

   ```
feat(cli): The `meltano run` command now accepts a `--timeout` option to
limit the time it runs
   ```

3. Provide a description of your changes.

4. Put "Closes #XXXX" in your comment to auto-close the issue that your
PR fixes (if such).

-->

## Description

<!-- Describe the changes introduced by this PR -->

## Related Issues

* Closes #XXXX

## Summary by Sourcery

Build:
- Restore support for Click 8.1 by adjusting the click dependency
constraint